Updated telegram model
